Steps on the way to 5.15.166

Resolves merge conflicts in:
	fs/fuse/inode.c

+++ include/linux/blkdev.h
@@ -239,7 +239,7 @@ struct request {
 	ANDROID_KABI_RESERVE(1);
 };
 
-static inline int blk_validate_block_size(unsigned int bsize)
+static inline int blk_validate_block_size(unsigned long bsize)
 {
 	if (bsize < 512 || bsize > PAGE_SIZE || !is_power_of_2(bsize))
 		return -EINVAL;
